Modifying block size to long long in do_IO

+++ utils.cc
@@ -127,7 +127,7 @@ inline unsigned long long get_file_size(string file_name){
 }
 
 int do_IO( string io_action, string path, int thread_count, int chunk_size,
-    	int block_size, int blocks_count ){
+    	unsigned long long block_size, int blocks_count ){
 
 	// Thread Identifiers for each thread that is going to be spawn
 	vector<pthread_t> threads(thread_count);
